# service/execution_openapi.yaml
openapi: 3.0.3
info:
  title: MT5 Demo Execution API
  version: 1.2.0
  description: Demo-only execution boundary. API keys authenticate but never authorize trading. Mutations require a locally provisioned durable scoped authorization bound to session and boot identity. Entry is closed by default; scoped exit recovery is independent of entry expiry or kill. Ambiguous outcomes must be reconciled, never automatically resent.
servers:
  - url: /api/v1
security:
  - ApiKeyAuth: []
paths:
  /symbols/{symbol}/quote:
    get:
      operationId: getExecutionQuote
      description: Direct broker quote via the same execution adapter and guarded Demo session. Compare fingerprint, generation and epoch with fresh health. fresh is false for stale or future quotes; never infer tradability from HTTP 200 alone.
      parameters:
        - $ref: '#/components/parameters/Symbol'
      responses:
        '200': {description: Timestamped bid/ask and logical-to-broker mapping, content: {application/json: {schema: {$ref: '#/components/schemas/QuoteRiskEnvelope'}}}}
        '401': {$ref: '#/components/responses/Error'}
        '503': {$ref: '#/components/responses/Error'}
  /symbols/{symbol}/risk:
    get:
      operationId: getExecutionRisk
      description: Quote plus loss-side tick value per lot, explicit account currency, BUY and SELL one-tick loss calculations via order_calc_profit, and actual fixed 20-point entry/close deviation. loss_tick_value is the maximum of trade_tick_value_loss and calculated losses, not general trade_tick_value. Point-in-time evidence, not a bound on gaps, commissions or future FX conversion. Missing calculations fail closed; stale quotes return risk_ready false.
      parameters:
        - $ref: '#/components/parameters/Symbol'
      responses:
        '200': {description: Account-currency loss and execution policy evidence, content: {application/json: {schema: {$ref: '#/components/schemas/RiskEnvelope'}}}}
        '401': {$ref: '#/components/responses/Error'}
        '503': {$ref: '#/components/responses/Error'}
...
  /account:
    get:
      operationId: getExecutionAccount
      responses:
        '200': {description: Opaque account identity and decimal balance facts, content: {application/json: {schema: {$ref: '#/components/schemas/AccountEnvelope'}}}}
        '401': {$ref: '#/components/responses/Error'}
        '503': {$ref: '#/components/responses/Error'}
        '500': {$ref: '#/components/responses/Error'}
  /symbols/{symbol}:
    get:
      operationId: getSymbolSpecification
      parameters:
        - $ref: '#/components/parameters/Symbol'
      responses:
        '200': {description: Fresh symbol and risk facts, content: {application/json: {schema: {$ref: '#/components/schemas/SymbolEnvelope'}}}}
        '400': {$ref: '#/components/responses/Error'}
        '401': {$ref: '#/components/responses/Error'}
        '404': {$ref: '#/components/responses/Error'}
        '503': {$ref: '#/components/responses/Error'}
        '500': {$ref: '#/components/responses/Error'}
  /orders:
    get:
      operationId: listOrders
      responses:
        '200': {description: Active and recent orders, content: {application/json: {schema: {$ref: '#/components/schemas/OrdersEnvelope'}}}}
        '401': {$ref: '#/components/responses/Error'}
        '503': {$ref: '#/components/responses/Error'}
        '500': {$ref: '#/components/responses/Error'}
    post:
      operationId: createOrder
      description: Accepts the bot OrderRequest and performs one order_check and at most one order_send. Only GTC market orders are supported; other order types or time-in-force values fail with 422 before send. MT5 timeout 10012 and connection loss 10031 remain indeterminate and require by-client recovery. Never automatically retry after timeout.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/RequestId'
        - $ref: '#/components/parameters/ExecutionAuthorization'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/CreateOrder'}
      responses:
        '200': {description: Original accepted result replay, content: {application/json: {schema: {$ref: '#/components/schemas/OrderResultEnvelope'}}}}
        '201': {description: Accepted by MT5, content: {application/json: {schema: {$ref: '#/components/schemas/OrderResultEnvelope'}}}}
        '400': {$ref: '#/components/responses/Error'}
        '401': {$ref: '#/components/responses/Error'}
        '403': {$ref: '#/components/responses/Error'}
        '409': {$ref: '#/components/responses/Error'}
        '422': {$ref: '#/components/responses/Error'}
        '503': {$ref: '#/components/responses/Error'}
        '500': {$ref: '#/components/responses/Error'}
  /orders/{id}:
    get:
      operationId: getOrder
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200': {description: Order, content: {application/json: {schema: {$ref: '#/components/schemas/OrderEnvelope'}}}}
        '400': {$ref: '#/components/responses/Error'}
        '401': {$ref: '#/components/responses/Error'}
        '404': {$ref: '#/components/responses/Error'}
        '503': {$ref: '#/components/responses/Error'}
        '500': {$ref: '#/components/responses/Error'}
  /orders/by-client/{client_order_id}:
    get:
      operationId: getOrderByClientId
      description: Durable timeout-after-acceptance recovery lookup. Returns an Order only after MT5 acceptance is observed on the original account and order facts match the original intent; 404 means acceptance is not yet known or rejection was observed. Legacy records lacking original account identity fail closed with 503. The durable record is retained and this endpoint never sends an order.
      parameters:
        - name: client_order_id
          in: path
          required: true
          schema: {type: string, minLength: 1}
      responses:
        '200': {description: Reconciled MT5 Order, content: {application/json: {schema: {$ref: '#/components/schemas/OrderEnvelope'}}}}
        '401': {$ref: '#/components/responses/Error'}
        '404': {$ref: '#/components/responses/Error'}
        '409': {$ref: '#/components/responses/Error'}
        '503': {$ref: '#/components/responses/Error'}
        '500': {$ref: '#/components/responses/Error'}
  /orders/{id}/cancel:
    post:
      operationId: cancelOrder
      description: Scoped recovery only for the verified authorized order. A repeated or ambiguous attempt never sends again. Entry kill does not revoke bounded exit recovery; session changes require operator escalation.
      parameters:
        - $ref: '#/components/parameters/Id'
        - $ref: '#/components/parameters/RequestId'
        - $ref: '#/components/parameters/ExecutionAuthorization'
      requestBody:
        required: false
        content:
          application/json:
            schema: {type: object, additionalProperties: false, maxProperties: 0}
      responses:
        '200': {description: Order observed after a single cancel attempt, content: {application/json: {schema: {$ref: '#/components/schemas/OrderEnvelope'}}}}
        '400': {$ref: '#/components/responses/Error'}
        '401': {$ref: '#/components/responses/Error'}
        '403': {$ref: '#/components/responses/Error'}
        '409': {$ref: '#/components/responses/Error'}
        '422': {$ref: '#/components/responses/Error'}
        '503': {$ref: '#/components/responses/Error'}
        '500': {$ref: '#/components/responses/Error'}
  /positions:
    get:
      operationId: listPositions
      responses:
        '200': {description: Open positions, content: {application/json: {schema: {$ref: '#/components/schemas/PositionsEnvelope'}}}}
        '401': {$ref: '#/components/responses/Error'}
        '503': {$ref: '#/components/responses/Error'}
        '500': {$ref: '#/components/responses/Error'}
  /positions/{id}/close:
    post:
      operationId: closePosition
      description: Scoped recovery only for exclusively owned authorized test exposure. Shared/netting exposure and changed sessions fail closed. Ambiguous outcomes require GET reconciliation and operator escalation, never automatic resend.
      parameters:
        - $ref: '#/components/parameters/Id'
        - $ref: '#/components/parameters/RequestId'
        - $ref: '#/components/parameters/ExecutionAuthorization'
      requestBody:
        required: false
        content:
          application/json:
            schema: {type: object, additionalProperties: false, maxProperties: 0}
      responses:
        '200': {description: Deal observed after a single close attempt, content: {application/json: {schema: {$ref: '#/components/schemas/DealEnvelope'}}}}
        '400': {$ref: '#/components/responses/Error'}
        '401': {$ref: '#/components/responses/Error'}
        '403': {$ref: '#/components/responses/Error'}
        '404': {$ref: '#/components/responses/Error'}
        '409': {$ref: '#/components/responses/Error'}
        '422': {$ref: '#/components/responses/Error'}
        '503': {$ref: '#/components/responses/Error'}
        '500': {$ref: '#/components/responses/Error'}
  /deals:
    get:
      operationId: listDeals
      responses:
        '200': {description: Recent trading deals; balance and other non-trading operations are excluded, content: {application/json: {schema: {$ref: '#/components/schemas/DealsEnvelope'}}}}
        '401': {$ref: '#/components/responses/Error'}
        '502': {$ref: '#/components/responses/Error'}
        '503': {$ref: '#/components/responses/Error'}
        '500': {$ref: '#/components/responses/Error'}
components:
  securitySchemes:
    ApiKeyAuth: {type: apiKey, in: header, name: X-API-Key}
  parameters:
    ExecutionAuthorization: {name: X-Execution-Authorization, in: header, required: true, description: ID of an immutable operator-local grant; API key alone cannot provision or enable a grant., schema: {type: string, minLength: 1}}
    Symbol: {name: symbol, in: path, required: true, schema: {type: string, minLength: 1}}
...
    IdempotencyKey: {name: Idempotency-Key, in: header, required: true, description: Must equal client_order_id., schema: {type: string, minLength: 1}}
    RequestId: {name: X-Request-ID, in: header, required: false, description: If present on submit, must equal OrderRequest.request_id., schema: {type: string, minLength: 1}}
  responses:
    Error:
      description: Stable versioned error
      headers:
        X-Request-ID: {schema: {type: string}, description: Correlates this response with the request.}
        X-MT5-Retcode: {schema: {type: integer}, description: Original MT5 retcode when available; kept outside the strict bot error DTO.}
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorEnvelope'}
  schemas:
    RiskEnvelope:
      allOf:
        - $ref: '#/components/schemas/QuoteRiskEnvelope'
        - type: object
          properties:
            data:
              type: object
              required: [loss_tick_value, tick_value_loss, loss_calculation_buy, loss_calculation_sell, tick_value_currency, account_currency, currency_relationship, loss_value_basis, tick_size, point, volume_min, volume_step, stops_level, freeze_level, entry_deviation_points, close_deviation_points, deviation_price, risk_ready]
    QuoteRiskEnvelope:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelope'
        - type: object
          properties:
            data:
              type: object
              required: [symbol, broker_symbol, bid, ask, quote_at, observed_at, age_seconds, max_age_seconds, fresh, account_session]
              properties:
                symbol: {type: string}
                broker_symbol: {type: string}
                bid: {$ref: '#/components/schemas/PositiveDecimal'}
                ask: {$ref: '#/components/schemas/PositiveDecimal'}
                quote_at: {type: string, format: date-time}
                observed_at: {type: string, format: date-time}
                age_seconds: {type: string}
                max_age_seconds: {type: integer, enum: [5]}
                fresh: {type: boolean}
                account_session:
                  type: object
                  required: [fingerprint, generation, epoch, account_mode, ready]
                  properties:
                    fingerprint: {type: string, minLength: 1}
                    generation: {type: integer, minimum: 1}
                    epoch: {type: string, minLength: 1}
                    account_mode: {type: string, enum: [DEMO]}
                    ready: {type: boolean, enum: [true]}
                loss_tick_value: {$ref: '#/components/schemas/PositiveDecimal'}
                tick_value_loss: {$ref: '#/components/schemas/PositiveDecimal'}
                loss_calculation_buy: {$ref: '#/components/schemas/PositiveDecimal'}
                loss_calculation_sell: {$ref: '#/components/schemas/PositiveDecimal'}
                tick_value_currency: {type: string}
                account_currency: {type: string}
                currency_relationship: {type: string}
                loss_value_basis: {type: string}
                tick_size: {$ref: '#/components/schemas/PositiveDecimal'}
                point: {$ref: '#/components/schemas/PositiveDecimal'}
                volume_min: {$ref: '#/components/schemas/PositiveDecimal'}
                volume_step: {$ref: '#/components/schemas/PositiveDecimal'}
                stops_level: {type: string}
                freeze_level: {type: string}
                entry_deviation_points: {type: integer, enum: [20]}
                close_deviation_points: {type: integer, enum: [20]}
                deviation_price: {$ref: '#/components/schemas/PositiveDecimal'}
                risk_ready: {type: boolean}
...
    SymbolEnvelope:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelope'
        - type: object
          properties:
            data: {$ref: '#/components/schemas/SymbolSpecification'}
...
    AccountEnvelope:
      allOf:
        - {$ref: '#/components/schemas/SuccessEnvelope'}
        - {type: object, properties: {data: {$ref: '#/components/schemas/AccountSnapshot'}}}
    OrdersEnvelope:
      allOf:
        - {$ref: '#/components/schemas/SuccessEnvelope'}
        - {type: object, properties: {data: {type: array, items: {$ref: '#/components/schemas/Order'}}}}
...